split off test projects
