+++ Cargo.toml
@@ -29,8 +29,12 @@ name = "oled-bus"
 path = "src/examples/ssd1306-oled-bus.rs"
 
 [[bin]]
-name = "test"
-path = "src/lib/test.rs"
+name = "signalk"
+path = "src/lib/test/signalk.rs"
+
+[[bin]]
+name = "constantsensor"
+path = "src/lib/test/constant_sensor.rs"
 
 [profile.release]
 opt-level = "s"
